+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.3.0 TBD
 - Change: A TUN-device is used instead of firewall rules for routing outbound connections.
 - Change: Outbound connections now use gRPC instead of ssh, and the traffic-manager no longer has a sshd running.
+- Change: The traffic-agent no longer has a sshd running. Remote volume mounts use sshfs in slave mode, talking directly to sftp. 
 
 ### 2.2.2 (May 17, 2021)
 - Feature: Telepresence translates legacy Telepresence commands into viable Telepresence commands.